+++ crates/holochain_launcher_utils/src/zome_call_signing.rs
@@ -2,7 +2,7 @@
 use holochain_client::AgentPubKey;
 use holochain_conductor_api::ZomeCall;
 use holochain_types::prelude::ZomeCallUnsigned;
-use holochain_zome_types::{Signature, CellId, ZomeName, FunctionName, CapSecret, ExternIO, Timestamp, Zome};
+use holochain_zome_types::{Signature, CellId, ZomeName, FunctionName, CapSecret, ExternIO, Timestamp};
 use lair_keystore_api::LairClient;
 
 use serde::Deserialize;
